Warnings resolved
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/mmstd_trisoup/trunk@194 cb50be13-817d-0410-ab5c-3777bfb244a9

+++ src/AbstractTriMeshDataSource.cpp
@@ -75,7 +75,7 @@ bool AbstractTriMeshDataSource::getDataCallback(core::Call& caller) {
         ctmd->SetUnlocker(NULL);
     } else if (ldc != NULL) {
         ldc->SetDataHash(this->datahash);
-        ldc->SetData(this->lines.size(), this->lines.data());
+        ldc->SetData(static_cast<unsigned int>(this->lines.size()), this->lines.data());
         ldc->SetUnlocker(NULL);
     }
